unsaved palette warning in all cases
